requirements.txt
googlesearch-python
beautifulsoup4
requests
openai
python-dotenv